Decipher messages encoded using a prefix code
Taking a string of 1's and 0's, build the prefix tree, traverse the tree
in preorder and prints the list of characetrs in the tree, the number of
bits, and their encoding. Uncompress (method) reads the compressed
message from the file and writes the uncompressed message using cout.
For example:

character encoding
a  0
b 111
c  1011
d 1010
r 110
! 100
Given the string: 0111110010110101001111100100
Run through the program yields:
"abracadabra!"