Encode messages using Huffman Coding
First, I mapped out each letter in the string to determine how many
times it appeared and pushed the resulting node onto a queue. Next,
using these nodes, I built a priority queue which was then used to make
the prefix tree (with lower numbered nodes on the top of the tree, going
to the highest).  Finally, I take the binary code from the tree, print
it, and store it using another map and writes this to a file. This new
file should be able to be run through the decoder (Decipher messages
encoded using a prefix tree) and vice versa.

Example:

Input File:
! => 1
a => 5
b => 2
c => 1
d => 1
r => 2

Total Characters: 12

Prefix Code Tree:
*a**db**!cr

Encodings:
a 0
d 100
b 101
! 1100
c 1101
r 111

Encoded message:
0101111011010100010111101100
Output file:
*a**db**!cr
101111011010100010111101100

***Note: CharCountNode.h was given by the instructor and was not altered
in any way***